+++ src/RealisableLeafFilterNodeExtensions.cs
@@ -5,19 +5,19 @@ namespace ExtremeAndy.CombinatoryFilters
     public static class RealisableLeafFilterNodeExtensions
     {
         [Obsolete("This method is comparatively slow and causes an allocation for each call. Use `GetPredicate` instead and save the result when testing many items.")]
-        public static bool IsMatch<TLeafNode, TItemToTest>(
-            this IFilterNode<TLeafNode> filter,
+        public static bool IsMatch<TFilter, TItemToTest>(
+            this IFilterNode<TFilter> filter,
             TItemToTest item)
-            where TLeafNode : class, ILeafFilterNode, IRealisableLeafFilterNode<TItemToTest>
+            where TFilter : IFilter<TItemToTest>
         {
-            var predicate = filter.GetPredicate<TLeafNode, TItemToTest>();
+            var predicate = filter.GetPredicate<TFilter, TItemToTest>();
             return predicate(item);
         }
 
-        public static Func<TItemToTest, bool> GetPredicate<TLeafNode, TItemToTest>(this IFilterNode<TLeafNode> filter)
-            where TLeafNode : class, ILeafFilterNode, IRealisableLeafFilterNode<TItemToTest>
+        public static Func<TItemToTest, bool> GetPredicate<TFilter, TItemToTest>(this IFilterNode<TFilter> filter)
+            where TFilter : IFilter<TItemToTest>
         {
-            return filter.GetPredicate<TLeafNode, TItemToTest>(leafFilter => leafFilter.IsMatch);
+            return filter.GetPredicate<TFilter, TItemToTest>(leafFilter => leafFilter.IsMatch);
         }
     }
 }