+++ tests/BindTests.cs
@@ -1,4 +1,3 @@
-using System.Collections.Generic;
 using Xunit;
 
 namespace ExtremeAndy.CombinatoryFilters.Tests
@@ -10,16 +9,16 @@ namespace ExtremeAndy.CombinatoryFilters.Tests
         {
             var filter5To10 = new NumericRangeFilter(5, 10);
             var filter8To15 = new NumericRangeFilter(8, 15);
-            var filter5To10Or8To15 = new CombinationFilter<NumericRangeFilter>(new[] { filter5To10, filter8To15 }, CombinationOperator.Any);
+            var filter5To10Or8To15 = new CombinationFilterNode<NumericRangeFilter>(new[] { filter5To10, filter8To15 }, CombinationOperator.Any);
             var filter9To12 = new NumericRangeFilter(9, 12);
-            var filter = new CombinationFilter<NumericRangeFilter>(new IFilterNode<NumericRangeFilter>[] { filter5To10Or8To15, filter9To12 }, CombinationOperator.All);
+            var filter = new CombinationFilterNode<NumericRangeFilter>(new IFilterNode<NumericRangeFilter>[] { filter5To10Or8To15, filter9To12.ToLeafFilterNode() }, CombinationOperator.All);
 
             var result = filter.Bind<NumericRangeFilter>(
                 f =>
                 {
                     if (ReferenceEquals(f, filter5To10))
                     {
-                        return new CombinationFilter<NumericRangeFilter>(new [] { filter5To10, filter8To15 }, CombinationOperator.Any);
+                        return new CombinationFilterNode<NumericRangeFilter>(new [] { filter5To10, filter8To15 }, CombinationOperator.Any);
                     }
 
                     if (ReferenceEquals(f, filter8To15))
@@ -29,25 +28,25 @@ namespace ExtremeAndy.CombinatoryFilters.Tests
 
                     if (ReferenceEquals(f, filter9To12))
                     {
-                        return new InvertedFilter<NumericRangeFilter>(f);
+                        return new InvertedFilterNode<NumericRangeFilter>(f);
                     }
 
-                    return f;
+                    return f.ToLeafFilterNode();
                 });
 
-            var expectedFilter = new CombinationFilter<NumericRangeFilter>(
+            var expectedFilter = new CombinationFilterNode<NumericRangeFilter>(
                 new IFilterNode<NumericRangeFilter>[]
                 {
-                    new CombinationFilter<NumericRangeFilter>(
+                    new CombinationFilterNode<NumericRangeFilter>(
                         new IFilterNode<NumericRangeFilter>[]
                         {
                             filter5To10Or8To15,
                             FilterNode<NumericRangeFilter>.False
                         }, CombinationOperator.Any),
-                    new InvertedFilter<NumericRangeFilter>(filter9To12),
+                    new InvertedFilterNode<NumericRangeFilter>(filter9To12),
                 }, CombinationOperator.All);
 
-            Assert.Equal(expectedFilter, result, EqualityComparer<IFilterNode>.Default);
+            Assert.Equal(expectedFilter, result);
         }
     }
 }