+++ tests/IsMatchTests.cs
@@ -10,9 +10,9 @@ namespace ExtremeAndy.CombinatoryFilters.Tests
         {
             var filter5To10 = new NumericRangeFilter(5, 10);
             var filter8To15 = new NumericRangeFilter(8, 15);
-            var filter5To10Or8To15 = new CombinationFilter<NumericRangeFilter>(new[] { filter5To10, filter8To15 }, CombinationOperator.Any);
+            var filter5To10Or8To15 = new CombinationFilterNode<NumericRangeFilter>(new[] { filter5To10, filter8To15 }, CombinationOperator.Any);
             var filter9To12 = new NumericRangeFilter(9, 12);
-            var filter = new CombinationFilter<NumericRangeFilter>(new IFilterNode<NumericRangeFilter>[] { filter5To10Or8To15, filter9To12 }, CombinationOperator.All);
+            var filter = new CombinationFilterNode<NumericRangeFilter>(new IFilterNode<NumericRangeFilter>[] { filter5To10Or8To15, filter9To12.ToLeafFilterNode() }, CombinationOperator.All);
 
             var values = new[] { 1, 3, 5, 9, 11 };
             var expectedFilteredValues = new[] { 9, 11 };